// GF2020-v1/src/app/models/order-management/river-bill-draft/rbl-fetch.ts
export interface RblFetchModel {
    customerimage: string;
    blnumber: string;
    mloblnumber: string;
    blCount: string;
    shippername: string;
    consigneename: string;
    commgrp: string;
    general: boolean;
    reefer: boolean;
    hazmat: boolean;
    empty: boolean;
    nhc: boolean;
    bldate: string;
    closedate: string;
    blstatus: string;
    cntrcnt: number;
    uploadCntr: number;
    cntrdisp: number;
    customername: string;
}