// GF2020-v1/src/app/models/order-management/river-bill-draft/save-rbl.ts
export interface SaveRbl {
    childbookingpk: number;
    locationfk: number;
    bookingmasterfk: number;
    bookingid: string;
    bookingdate: string;
    referencetype: number;
    referencefk: number;
    referencenumber: string;
    customerfk: number;
    shipperfk: number;
    consigneefk: number;
    polfk: number;
    polterfk: number;
    polagentfk: number;
    podfk: number;
    podterfk: number;
    podagentfk: number;
    tosfk: number;
    bookingtype: number;
    genteus: number;
    reeferteus: number;
    hazmatteus: number;
    ladenwgt: number;
    totalwgt: number;
    emptyteus: number;
    emptywgt: number;
    totalteus: number;
    bookingstatus: string;
    mloblnr: string;
    bkgcur: number;
    bkgamt: number;
    is_active: number;
    created_by_fk: number;
    created_on: string;
    last_updated_by_fk: number;
    last_updated_on: string;
    version_no: number;
    commercial_schedule_hdr_fk: number;
    blnumber: string;
    bldate: string;
    blstatus: number;
    priority: number;
    notify_fk: number;
    shippername: string;
    shipperadd: string;
    shipperadd2: string;
    shippercity: string;
    shippercountry: string;
    shipperrucnr: string;
    consgineename: string;
    consigneeadd: string;
    consigneeadd2: string;
    consigneecity: string;
    consigneecountry: string;
    consigneerucnr: string;
    notifyname: string;
    notifyadd: string;
    notifyadd2: string;
    notifycity: string;
    notifycountry: string;
    notifyrucnr: string;
    cntrcnt: number;
    despachodenr: string;
    cntrList: [
        {
            childbkgcontainerpk: number;
            childbookingfk: number;
            childbkgctypefk: number;
            bookingcontainerfk: number;
            shipper: string;
            shipperadd: string;
            consignee: string;
            consigneeadd: string;
            notifyparty: string;
            notifyadd: string;
            mothervessel: string;
            mlo: string;
            mloblnr: string;
            childbkgnr: string;
            containernumber: string;
            containertypefk: number;
            commgrpfk: number;
            commfk: number;
            grossweight: number;
            tareweight: number;
            volume: number;
            imdg: string;
            unnr: string;
            temp: string;
            containerseal: string;
            customsseal: string;
            packtype: string;
            packqty: number;
            marks: string;
            description: string;
            ncmnr: string;
            remarks1: string;
            remarks2: string;
            remarks3: string;
            is_active: number;
            created_by_fk: number;
            created_on: string;
            last_updated_by_fk: number;
            last_updated_on: string;
            version_no: number;
            hazmat_approval: number;
            reefer_approval: number;
            nch_file_data: string;
            containersize: string;
            loaded: boolean;
            load_date: string;
            load_print: string;
            discharged: boolean;
            discharge_date: string;
            discharge_print: string;
            dischargetime: string;
            loadtime: string;
            stowpos: string;
            haz_ref_app_date: string;
            shipperadd2: string;
            shippercity: string;
            shippercountry: string;
            shipperrucnr: string;
            consigneeadd2: string;
            consigneecity: string;
            consigneecountry: string;
            consigneerucnr: string;
            notifyadd2: string;
            notifycity: string;
            notifycountry: string;
            notifyrucnr: string
        }
    ]
}