// GF2020-v1/src/app/models/order-management/river-bill-draft/fetch-rbl-entry.ts
export interface FetchRblEntryModel {
    childbookingpk: number;
    locationfk: number;
    bookingid: string;
    bookingmasterfk: number;
    bookingdate: number;
    referencetype: string;
    referencefk: number;
    referencenumber: string;
    customerfk: number;
    shipperfk: string;
    consigneefk: string;
    polfk: number;
    polterfk: number;
    polagentfk: number;
    podfk: number;
    podterfk: number;
    podagentfk: number;
    tosfk: string;
    bookingtype: number;
    ladenwgt: number;
    bookingstatus: string;
    mloblnr: string;
    bkgcur: number;
    bkgamt: number;
    is_active: number;
    created_by_fk: number;
    created_on: string;
    last_updated_by_fk: string;
    last_updated_on: string;
    version_no: string;
    commercial_schedule_hdr_fk: number;
    blnumber: string;
    bldate: string;
    blstatus: number;
    priority: string;
    notify_fk: string;
    shippername: string;
    shipperadd: string;
    shipperadd2: string;
    shippercity: string;
    shippercountry: string;
    shipperrucnr: string;
    consgineename: string;
    consigneeadd: string;
    consigneeadd2: string;
    consigneecity: string;
    consigneecountry: string;
    consigneerucnr: string;
    notifyname: string;
    notifyadd: string;
    notifyadd2: string;
    notifycity: string;
    notifycountry: string;
    notifyrucnr: string;
    cntrcnt: string;
    despachodenr: string;
    containernumber: string;
    customercode: string;
    customername: string;
    customerimage: string;
    servicecode: string;
    servicename: string;
    voynr: string;
    poltername: string;
    poleta: string;
    poletd: string;
    podtername: string;
    podeta: string;
    podetd: string;
    blstatusdesc: string;
    totalteus: number;
    genteus: number;
    reeferteus: number;
    hazmatteus: number;
    emptyteus: number;
    bkgWgt: number;
    genWgt: number;
    reeferWgt: number;
    hazWgt: number;
    emptywgt: number;
    regdesalidat: string;
    polname: string;
    podname: string;
    loaddate: string;
    sh_cntry_name: number;
    cn_cntry_name: number;
    np_cntry_name: number;
    sameasconsignee: true;
    cntrlist: [
        {
            childbkgcontainerpk: number;
            childbookingfk: number;
            childbkgctypefk: number;
            bookingcontainerfk: number;
            shipper: string;
            shipperadd: string;
            consignee: string;
            consigneeadd: string;
            notifyparty: string;
            notifyadd: string;
            mothervessel: string;
            mlo: string;
            mloblnr: string;
            childbkgnr: string;
            containernumber: string;
            containertypefk?: number;
            commgrpfk?: number;
            commfk?: number;
            grossweight: number;
            tareweight: number;
            volume: number;
            imdg: string;
            unnr: string;
            temp: string;
            containerseal: string;
            customsseal: string;
            packtype: string;
            packqty: number;
            marks: string;
            description: string;
            ncmnr: string;
            remarks1: string;
            remarks2: string;
            remarks3: string;
            is_active: number;
            created_by_fk: number;
            created_on: string;
            last_updated_by_fk: number;
            last_updated_on: string;
            version_no: string;
            hazmat_approval: number;
            reefer_approval: number;
            nch_file_data: string;
            containersize: string;
            loaded: false;
            load_date: string;
            load_print: string;
            discharged: false;
            discharge_date: string;
            discharge_print: string;
            dischargetime: string;
            loadtime: string;
            stowpos: string;
            haz_ref_app_date: string;
            shipperadd2: string;
            shippercity: string;
            shippercountry: string;
            shipperrucnr: string;
            consigneeadd2: string;
            consigneecity: string;
            consigneecountry: string;
            consigneerucnr: string;
            notifyadd2: string;
            notifycity: string;
            notifycountry: string;
            notifyrucnr: string;
            is_lcl: number;
            containertype_id: string;
            cntrtype: string;
            netweight: number;
            commoditygroup_pk?: number;
            commoditygroup_id: string;
            commodityname: string;
            cntrpref: number
        }
    ];
}